+++ hc/front/views.py
@@ -174,10 +174,15 @@ def my_checks(request, code):
         request.profile.sort = request.GET["sort"]
         request.profile.save()
 
+    if request.GET.get("urls") in ("uuid", "slug") and rw:
+        project.show_slugs = request.GET["urls"] == "slug"
+        project.save()
+
     if request.session.get("last_project_id") != project.id:
         request.session["last_project_id"] = project.id
 
     q = Check.objects.filter(project=project)
+    q = q.select_related("project")
     checks = list(q.prefetch_related("channel_set"))
     sortchecks(checks, request.profile.sort)
 
@@ -204,6 +209,15 @@ def my_checks(request, code):
             if search not in search_key:
                 hidden_checks.add(check)
 
+    # Figure out which checks have ambiguous ping URLs
+    seen, ambiguous = set(), set()
+    if project.show_slugs:
+        for check in checks:
+            if check.slug and check.slug in seen:
+                ambiguous.add(check.slug)
+            else:
+                seen.add(check.slug)
+
     # Do we need to show the "Last Duration" header?
     show_last_duration = False
     for check in checks:
@@ -226,6 +240,7 @@ def my_checks(request, code):
         "selected_tags": selected_tags,
         "search": search,
         "hidden_checks": hidden_checks,
+        "ambiguous": ambiguous,
         "show_last_duration": show_last_duration,
     }
 
@@ -402,7 +417,7 @@ def update_name(request, code):
 
     form = forms.NameTagsForm(request.POST)
     if form.is_valid():
-        check.name = form.cleaned_data["name"]
+        check.set_name_slug(form.cleaned_data["name"])
         check.tags = form.cleaned_data["tags"]
         check.desc = form.cleaned_data["desc"]
         check.save()
@@ -694,7 +709,7 @@ def copy(request, code):
         new_name = check.name[:90] + "... (copy)"
 
     copied = Check(project=check.project)
-    copied.name = new_name
+    copied.set_name_slug(new_name)
     copied.desc, copied.tags = check.desc, check.tags
     copied.subject, copied.subject_fail = check.subject, check.subject_fail
     copied.methods = check.methods