+++ CHANGELOG.md
@@ -6,11 +6,15 @@ All notable changes to this project will be documented in this file.
 ### Improvements
 - Add /api/v1/badges/ endpoint (#552)
 - Add ability to edit existing email, Signal, SMS, WhatsApp integrations
+- Add new ping URL format: /{ping_key}/{slug} (#491)
 
 ### Bug Fixes
 - Add handling for non-latin-1 characters in webhook headers
 - Fix dark mode bug in selectpicker widgets
 - Fix a crash during login when user's profile does not exist (#77)
+- Drop API support for GET, DELETE requests with a request body
+- Add missing @csrf_exempt annotations in API views
+- Fix the ping handler to reject status codes > 255
 
 ## v1.22.0 - 2020-08-06